+++ brewtils/plugin.py
@@ -51,6 +51,8 @@ request_context.current_request = None
 
 # Global config, used to simplify BG client creation and sanity checks.
 CONFIG = Box(default_box=True)
+# Global client
+CLIENT = None
 
 
 def get_current_request_read_only():
@@ -294,7 +296,6 @@ class Plugin(object):
         # Now set up the system
         self._system = self._setup_system(system, kwargs)
 
-        global CLIENT
         # Make sure this is set after self._system
         if client:
             self._set_client(client)
@@ -438,7 +439,8 @@ class Plugin(object):
         client_clazz._current_request = client_clazz.current_request
 
         self._client = new_client
-        brewtils.plugin.CLIENT = new_client
+        global CLIENT
+        CLIENT = new_client
 
     @property
     def system(self):