+++ CHANGELOG.rst
@@ -5,8 +5,8 @@ Brewtils Changelog
 ------
 TBD
 
+- Fixed issue trying to check the bytes hash in resolver 
 - Added `is_newer` support for various models to enable improved event handling in framework 
-
 
 3.29.1
 ------