+++ CHANGELOG.rst
@@ -7,10 +7,9 @@ TBD
 
 - Fixed issue trying to check the bytes hash in resolver 
 - Added `is_newer` support for various models to enable improved event handling in framework
-- Updated StatusInfo heartbeat to append no more than one 'NOT_CONFIGURED' to history
-- Added `is_newer` support for various models to enable improved event handling in framework 
 - Added reattempt for file chunk download
 - Added Entry Heartbeat event label
+- Updated StatusInfo heartbeat to append no more than one 'NOT_CONFIGURED' to history
 
 3.29.1
 ------